Moved and renamed UpgradeSupportStyles.ts to LinkStyles as it is being used on different journeys
