+++ client/components/mma/upgrade/UpgradeSupportThankYou.tsx
@@ -26,6 +26,7 @@ import {
 	listWithDividersCss,
 	whatHappensNextIconCss,
 } from '../../../styles/GenericStyles';
+import { linkCss } from '../../../styles/LinkStyles';
 import {
 	signInContentContainerCss,
 	signInCss,
@@ -38,7 +39,6 @@ import type {
 	UpgradeSupportInterface,
 } from './UpgradeSupportContainer';
 import { UpgradeSupportContext } from './UpgradeSupportContainer';
-import { linkCss } from './UpgradeSupportStyles';
 
 export const UpgradeSupportThankYou = () => {
 	const upgradeSupportContext = useContext(
